# src/product-service/docs/swagger.yaml
definitions:
  common.AppError:
    properties:
      key:
        type: string
      log:
        type: string
      message:
        type: string
      status_code:
        type: integer
    type: object
  common.ErrorResponse:
    properties:
      message:
        type: string
    type: object
  common.Response:
    properties:
      code:
        type: integer
      data: {}
      filter: {}
      message: {}
      paging: {}
    type: object
...
  models.AddSupplier:
    properties:
      address:
        type: string
      city:
        type: string
      country:
        type: string
      email:
        type: string
      phone:
        type: string
      supplier_name:
        type: string
    type: object
...
  models.InventoryTransaction:
    properties:
      product_id:
        type: string
      quantity:
        type: integer
      transaction_date:
        type: string
      transaction_id:
        type: string
      transaction_type:
        $ref: '#/definitions/models.Type'
      warehouse_id:
        type: integer
    type: object
...
  models.Type:
    enum:
    - 0
    - 1
    type: integer
    x-enum-comments:
      In: "0"
      Out: "1"
    x-enum-varnames:
    - In
    - Out
  models.UpdateSupplier:
    properties:
      address:
        type: string
      city:
        type: string
      country:
        type: string
      email:
        type: string
      phone:
        type: string
      supplier_id:
        type: integer
      supplier_name:
        type: string
    type: object
...
info:
  contact: {}
paths:
...
  /api/supplier:
    post:
      consumes:
      - application/json
      description: Create a new supplier with the provided parameters
      parameters:
      - description: Supplier to create
        in: body
        name: Supplier
        required: true
        schema:
          $ref: '#/definitions/models.AddSupplier'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/common.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/common.AppError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/common.AppError'
      summary: Create a new Supplier
      tags:
      - suppliers
  /api/supplier/{id}:
    delete:
      consumes:
      - application/json
      description: Delete a Supplier by id
      parameters:
      - description: Supplier ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/common.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/common.AppError'
      summary: Deleted a Supplier
      tags:
      - suppliers
    get:
      consumes:
      - application/json
      description: Get a Supplier by id
      parameters:
      - description: Supplier ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/common.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/common.AppError'
      summary: Get a Supplier
      tags:
      - suppliers
    put:
      consumes:
      - application/json
      description: Update an existing Supplier with the provided parameters
      parameters:
      - description: Supplier to update
        in: body
        name: Supplier
        required: true
        schema:
          $ref: '#/definitions/models.UpdateSupplier'
      - description: Supplier ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/common.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/common.AppError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/common.AppError'
      summary: Update an existing Supplier
      tags:
      - suppliers
  /api/suppliers:
    get:
      consumes:
      - application/json
      description: Retrieve all products, with optional paging
      parameters:
      - in: query
        name: limit
        type: integer
      - in: query
        name: page
        type: integer
      - in: query
        name: '-'
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/common.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/common.AppError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/common.AppError'
      summary: Retrieve a list of Suppliers
      tags:
      - suppliers
  /api/transaction:
    post:
      consumes:
      - application/json
      description: Create a new inventory  with the provided parameters
      parameters:
      - description: Inventory to create
        in: body
        name: inventoryData
        required: true
        schema:
          $ref: '#/definitions/models.InventoryTransaction'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/common.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/common.AppError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/common.ErrorResponse'
      summary: Create a new inventory
      tags:
      - transactions
  /api/transaction/{id}:
    delete:
      consumes:
      - application/json
      description: Delete a inventory transaction by id
      parameters:
      - description: Transaction ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/common.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/common.ErrorResponse'
      summary: Deleted a inventory transaction
      tags:
      - transactions
  /api/transactions:
    get:
      consumes:
      - application/json
      description: Retrieve all inventory , with optional paging
      parameters:
      - in: query
        name: limit
        type: integer
      - in: query
        name: page
        type: integer
      - in: query
        name: '-'
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/common.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/common.AppError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/common.ErrorResponse'
      summary: Retrieve a list of inventory
      tags:
      - transactions
...
swagger: "2.0"